Extract Failures to a separate file
